tiny: cast ino_t at printing
ino_t may be 32/64 bits.

Acked-by: Ryousei Takano <takano-ryousei@aist.go.jp>

+++ plink.c
@@ -289,7 +289,7 @@ static int do_plink(char *cwd, int cmd, int nbr, char *br[])
 	if (cmd == AuPlink_LIST) {
 		ia.p = ia.o;
 		for (i = 0; i < ia.nino; i++)
-			printf("%llu ", *ia.cur++);
+			printf("%llu ", (unsigned long long)*ia.cur++);
 		putchar('\n');
 	}
